==> 1_check.R <==
library(tidyverse)
library(vcfR)
library(do)
library(R.utils)
library(tidyverse)
library(multtest)
library(gplots)
library(LDheatmap)
library(genetics)
library(ape)
library(EMMREML)
library(Matrix)
library(compiler) 
library(scatterplot3d)
library(grid)
source("./01_scripts/GAPIT1.txt")
source("./01_scripts/GAPIT2.txt")
list_gene <- list.files("./02_ordata/",pattern = "*.vcf.gz")
trait <- read_table(paste0("05_trait/","trait.txt"),show_col_types = F)
print("################## check OK! #######################")
list_gene <- list.files("./02_ordata/",pattern = "*.vcf.gz")
list_gene <- substr(list_gene,1,19)
print("您当前处理的基因型文件如下：")
print(list_gene)
write.table(list_gene,"./01_scripts/list_gene.txt",sep = "\t",row.names = F,col.names = F,quote = F)


list_phe <- colnames(trait)
list_phe <- list_phe[-1]
print("您当前处理的表型型文件如下：")
print(list_phe)
write.table(list_phe,"./01_scripts/list_phe.txt",sep = "\t",row.names = F,col.names = F,quote = F)
print("#####################################################")
print("请检查是否有误，脚本即将自动运行......[ 5 秒后自动开始单基因关联分析]")
